refactor(FileLoader): move to typescript
